Fixing query to fetch listof values category by service id acc to changed req
