+++ src/functionalTest/java/uk/gov/hmcts/reform/cdapi/CommonDataApiFunctionalTest.java
@@ -1,266 +1,33 @@
 package uk.gov.hmcts.reform.cdapi;
 
 import io.restassured.response.Response;
-import org.junit.Ignore;
+import lombok.extern.slf4j.Slf4j;
+import net.thucydides.core.annotations.WithTag;
+import net.thucydides.core.annotations.WithTags;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.http.HttpStatus;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
 import uk.gov.hmcts.reform.cdapi.controllers.response.Categories;
-import uk.gov.hmcts.reform.cdapi.domain.CaseFlag;
-import uk.gov.hmcts.reform.cdapi.exception.ErrorResponse;
-import uk.gov.hmcts.reform.cdapi.util.ErrorInvalidRequestResponse;
 import uk.gov.hmcts.reform.cdapi.util.FeatureToggleConditionExtension;
 import uk.gov.hmcts.reform.cdapi.util.ToggleEnable;
+import uk.gov.hmcts.reform.lib.util.serenity5.SerenityTest;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.springframework.http.HttpStatus.NOT_FOUND;
 import static org.springframework.http.HttpStatus.OK;
 
-@Ignore
+@SerenityTest
+@SpringBootTest
+@WithTags({@WithTag("testType:Functional")})
+@ActiveProfiles("functional")
+@Slf4j
 class CommonDataApiFunctionalTest extends AuthorizationFunctionalTest {
 
-    private static final String MAP_KEY_CASE_FLAGS = "CaseFlagApiController.retrieveCaseFlagsByServiceId";
-    private static final String SLASH = "/";
-    private static final String PATH_CASE_FLAGS = SLASH.concat("caseflags");
     private static final String MAP_KEY_LOV = "CrdApiController.retrieveListOfValuesByCategoryId";
+    private static final String SLASH = "/";
     private static final String PATH_LOV = SLASH.concat("lov").concat(SLASH).concat("categories");
-    private static final String PARAM_SIGN = "?";
-    private static final String PARAM_HEARING = SLASH.concat("HearingChannel");
-    private static final String PARAM_HEARING_WITH_PARAM_SIGN = PARAM_HEARING.concat(PARAM_SIGN);
-    private static final String SERVICE_ID = "ServiceId=";
-    private static final String SERVICE_ID_BBA3 = "BBA3";
-    private static final String PARAM_LISTING_STATUS_WITH_SERVICE_ID = SLASH.concat("ListingStatus")
-        .concat(PARAM_SIGN).concat(SERVICE_ID + SERVICE_ID_BBA3);
-    private static final String PARAM_LISTING_STATUS_WITH_EMPTY_SERVICE_ID = SLASH.concat("ListingStatus")
-        .concat(PARAM_SIGN).concat(SERVICE_ID);
-    private static final String DATA_NOT_FOUND = "Data not found";
-
-    @Test
-    @ExtendWith(FeatureToggleConditionExtension.class)
-    @ToggleEnable(mapKey = MAP_KEY_CASE_FLAGS, withFeature = true)
-    void shouldThrowErrorWhenFlagTypeIsInvalid() {
-        final var response = (ErrorResponse)
-            commonDataApiClient.retrieveCaseFlagsByServiceId(
-                HttpStatus.BAD_REQUEST,
-                "service-id=XXXX?flag-type=hello"
-            );
-        assertNotNull(response);
-        assertEquals("Allowed values are PARTY or CASE", response.getErrorDescription());
-    }
-
-    @Test
-    @ExtendWith(FeatureToggleConditionExtension.class)
-    @ToggleEnable(mapKey = MAP_KEY_CASE_FLAGS, withFeature = false)
-    void shouldNotRetrieveCaseFlagWhenToggleOffWithStatusCode403() {
-        ErrorResponse response = (ErrorResponse)
-            commonDataApiClient
-                .retrieveResponseForGivenRequest(HttpStatus.FORBIDDEN,
-                                                 "/service-id=xxxx?flag-type=hello",
-                                                 CaseFlag.class, PATH_CASE_FLAGS
-                );
-        assertNotNull(response);
-    }
-
-    @Test
-    @ToggleEnable(mapKey = MAP_KEY_CASE_FLAGS, withFeature = true)
-    void retrieveCaseFlagsUnauthorizedDueToNoBearerTokenShouldReturnStatusCode401() {
-        Response response =
-            commonDataApiClient.retrieveResponseForGivenRequest_NoBearerToken("/service-id=AAA1", PATH_CASE_FLAGS);
-
-        assertNotNull(response);
-        assertThat(response.getHeader("UnAuthorized-Token-Error").contains("Authentication Exception"));
-        assertEquals(HttpStatus.UNAUTHORIZED.value(), response.getStatusCode());
-    }
-
-    @Test
-    @ToggleEnable(mapKey = MAP_KEY_CASE_FLAGS, withFeature = true)
-    void retrieveBuildingLocationsUnauthorizedDueToNoS2STokenShouldReturnStatusCode401() {
-        Response response =
-            commonDataApiClient.retrieveResponseForGivenRequest_NoS2SToken("/service-id=AAA1", PATH_CASE_FLAGS);
-
-        assertNotNull(response);
-        assertEquals(HttpStatus.UNAUTHORIZED.value(), response.getStatusCode());
-    }
-
-    @Test
-    @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
-    @ExtendWith(FeatureToggleConditionExtension.class)
-    void shouldReturnEmptyListWhenNoDataFound() {
-        final var response = (ErrorResponse)
-            commonDataApiClient.retrieveListOfValuesByCategoryId(
-                NOT_FOUND,
-                "hello"
-            );
-        assertNotNull(response);
-        assertEquals(DATA_NOT_FOUND, response.getErrorDescription());
-    }
-
-    @Test
-    @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
-    @ExtendWith(FeatureToggleConditionExtension.class)
-    void shouldReturnSuccess() {
-        Response response =
-            commonDataApiClient.retrieveCategoriesByCategoryIdSuccess(PATH_LOV, PARAM_HEARING);
-
-        if (OK.value() == response.getStatusCode()) {
-            var categories = response.getBody().as(Categories.class);
-            assertNotNull(categories);
-            assertThat(categories.getListOfCategory()).hasSizeGreaterThan(1);
-            categories.getListOfCategory().forEach(h -> assertEquals("HearingChannel", h.getCategoryKey()));
-            categories.getListOfCategory().forEach(h -> assertNull(h.getChildNodes()));
-        } else {
-            assertEquals(NOT_FOUND.value(), response.getStatusCode());
-        }
-
-    }
-
-    @Test
-    @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
-    @ExtendWith(FeatureToggleConditionExtension.class)
-    void shouldReturnSuccessWithChildren() {
-        Response response =
-            commonDataApiClient.retrieveCategoriesByCategoryIdSuccess(PATH_LOV, PARAM_HEARING_WITH_PARAM_SIGN
-                + "isChildRequired=y&key=telephone");
-
-        if (OK.value() == response.getStatusCode()) {
-            var categories = response.getBody().as(Categories.class);
-            assertNotNull(categories);
-            assertThat(categories.getListOfCategory()).hasSizeGreaterThan(0);
-            categories.getListOfCategory().forEach(h -> assertEquals("HearingChannel", h.getCategoryKey()));
-            categories.getListOfCategory().forEach(h -> assertFalse(h.getChildNodes().isEmpty()));
-        } else {
-            assertEquals(NOT_FOUND.value(), response.getStatusCode());
-        }
-    }
-
-    @Test
-    @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
-    @ExtendWith(FeatureToggleConditionExtension.class)
-    void shouldReturnSuccessForChildCategories() {
-        Response response =
-            commonDataApiClient.retrieveCategoriesByCategoryIdSuccess(PATH_LOV, "/HearingSubChannel?"
-                + "isChildRequired=y&parentKey=telephone");
-
-        if (OK.value() == response.getStatusCode()) {
-            var categories = response.getBody().as(Categories.class);
-            assertNotNull(categories);
-            categories.getListOfCategory().forEach(h -> assertEquals("HearingSubChannel", h.getCategoryKey()));
-            assertThat(categories.getListOfCategory()).hasSizeGreaterThan(0);
-        } else {
-            assertEquals(NOT_FOUND.value(), response.getStatusCode());
-        }
-    }
-
-    @Test
-    @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
-    @ExtendWith(FeatureToggleConditionExtension.class)
-    void shouldReturnSuccessWithNoChilds() {
-        Response response =
-            commonDataApiClient.retrieveCategoriesByCategoryIdSuccess(PATH_LOV, PARAM_HEARING_WITH_PARAM_SIGN
-                + "isChildRequired=n&key=telephone");
-
-        if (OK.value() == response.getStatusCode()) {
-            var categories = response.getBody().as(Categories.class);
-            assertNotNull(categories);
-            categories.getListOfCategory().forEach(h -> assertEquals("HearingChannel", h.getCategoryKey()));
-            assertThat(categories.getListOfCategory()).hasSizeGreaterThan(0);
-        } else {
-            assertEquals(NOT_FOUND.value(), response.getStatusCode());
-        }
-    }
-
-    @Test
-    @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
-    @ExtendWith(FeatureToggleConditionExtension.class)
-    void shouldReturnResourceNotFoundWithChildsAndInvalidParentKey() {
-        final var response = (ErrorResponse)
-            commonDataApiClient.retrieveListOfValuesByCategoryId(
-                NOT_FOUND,
-                PARAM_HEARING_WITH_PARAM_SIGN
-                    + "isChildRequired=y&parentKey=telephone"
-            );
-        assertNotNull(response);
-        assertEquals(DATA_NOT_FOUND, response.getErrorDescription());
-    }
-
-    @Test
-    @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
-    @ExtendWith(FeatureToggleConditionExtension.class)
-    void shouldReturnResourceNotFoundWithChildsAndInvalidParentCategory() {
-        final var response = (ErrorResponse)
-            commonDataApiClient.retrieveListOfValuesByCategoryId(
-                NOT_FOUND,
-                PARAM_HEARING_WITH_PARAM_SIGN
-                    + "isChildRequired=y&parentCategory=HearingChannel"
-            );
-        assertNotNull(response);
-        assertEquals(DATA_NOT_FOUND, response.getErrorDescription());
-    }
-
-    @Test
-    @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
-    @ExtendWith(FeatureToggleConditionExtension.class)
-    void shouldReturnResourceNotFoundInvalidServiceId() {
-        final var response = (ErrorResponse)
-            commonDataApiClient.retrieveListOfValuesByCategoryId(
-                NOT_FOUND,
-                PARAM_HEARING_WITH_PARAM_SIGN
-                    + "serviceId=BBA5"
-            );
-        assertNotNull(response);
-        assertEquals(DATA_NOT_FOUND, response.getErrorDescription());
-    }
-
-    @Test
-    @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
-    @ExtendWith(FeatureToggleConditionExtension.class)
-    void shouldReturnResourceNotFoundWithChildrenAndInvalidKey() {
-        final var response = (ErrorResponse)
-            commonDataApiClient.retrieveListOfValuesByCategoryId(
-                NOT_FOUND,
-                PARAM_HEARING_WITH_PARAM_SIGN
-                    + "isChildRequired=y&key=telephone-CVP"
-            );
-        assertNotNull(response);
-        assertEquals(DATA_NOT_FOUND, response.getErrorDescription());
-    }
-
-    @Test
-    @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
-    void retrieveCategoriesUnauthorizedDueToNoBearerTokenShouldReturnStatusCode401() {
-        Response response =
-            commonDataApiClient.retrieveResponseForGivenRequest_NoBearerToken(PARAM_HEARING, PATH_LOV);
-
-        assertNotNull(response);
-        assertEquals(HttpStatus.UNAUTHORIZED.value(), response.getStatusCode());
-    }
-
-    @Test
-    @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
-    void retrieveCategoriesUnauthorizedDueToNoS2STokenShouldReturnStatusCode401() {
-        Response response =
-            commonDataApiClient.retrieveResponseForGivenRequest_NoS2SToken(PARAM_HEARING, PATH_LOV);
-
-        assertNotNull(response);
-        assertEquals(HttpStatus.UNAUTHORIZED.value(), response.getStatusCode());
-    }
-
-    @Test
-    @ExtendWith(FeatureToggleConditionExtension.class)
-    @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
-    void shouldThrowErrorWhenCategoryIdIsEmpty() {
-        final var response = (ErrorInvalidRequestResponse)
-            commonDataApiClient.retrieveBadRequestByEmptyCategoryId(
-                NOT_FOUND,
-                ""
-            );
-        assertNotNull(response);
-        assertEquals("Not Found", response.getError());
-    }
 
     @Test
     @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
@@ -274,6 +41,8 @@ class CommonDataApiFunctionalTest extends AuthorizationFunctionalTest {
             var categories = response.getBody().as(Categories.class);
             assertNotNull(categories);
             categories.getListOfCategory().forEach(h -> {
+                log.info("***************** h:: " + h.getCategoryKey() + h.getServiceId()
+                             + h.getChildNodes().size() + h.getKey());
                 assertEquals("HearingChannel", h.getCategoryKey());
             });
             //assertThat(categories.getListOfCategory().size()).isGreaterThan(0);
@@ -287,71 +56,6 @@ class CommonDataApiFunctionalTest extends AuthorizationFunctionalTest {
         }
     }
 
-    @Test
-    @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
-    @ExtendWith(FeatureToggleConditionExtension.class)
-    void shouldReturnSuccessWithServiceIdNoChilds() {
-        Response response =
-            commonDataApiClient.retrieveCategoriesByCategoryIdSuccess(PATH_LOV, PARAM_HEARING_WITH_PARAM_SIGN
-                + "isChildRequired=n&serviceId=BBA3");
 
-        if (OK.value() == response.getStatusCode()) {
-            var categories = response.getBody().as(Categories.class);
-            assertNotNull(categories);
-            categories.getListOfCategory().forEach(h -> assertEquals("HearingChannel", h.getCategoryKey()));
-            assertThat(categories.getListOfCategory()).hasSizeGreaterThan(0);
-            categories.getListOfCategory().forEach(h -> assertNull(h.getChildNodes()));
-        } else {
-            assertEquals(NOT_FOUND.value(), response.getStatusCode());
-        }
-    }
-
-    @Test
-    @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
-    @ExtendWith(FeatureToggleConditionExtension.class)
-    void shouldReturnSuccess_Valid_ServiceID() {
-        Response response =
-            commonDataApiClient.retrieveCategoriesByCategoryIdSuccess(PATH_LOV, PARAM_LISTING_STATUS_WITH_SERVICE_ID);
-        if (OK.value() == response.getStatusCode()) {
-            var categories = response.getBody().as(Categories.class);
-            assertNotNull(categories);
-            assertThat(categories.getListOfCategory()).hasSizeGreaterThan(1);
-            categories.getListOfCategory().forEach(h -> assertEquals("ListingStatus", h.getCategoryKey()));
-        } else {
-            assertEquals(NOT_FOUND.value(), response.getStatusCode());
-        }
-    }
-
-    @Test
-    @ToggleEnable(mapKey = MAP_KEY_LOV, withFeature = true)
-    @ExtendWith(FeatureToggleConditionExtension.class)
-    void shouldReturnSuccess_Empty_ServiceID() {
-        Response response =
-            commonDataApiClient.retrieveCategoriesByCategoryIdSuccess(
-                PATH_LOV,
-                PARAM_LISTING_STATUS_WITH_EMPTY_SERVICE_ID
-            );
-        if (OK.value() == response.getStatusCode()) {
-            var categories = response.getBody().as(Categories.class);
-            assertNotNull(categories);
-            assertThat(categories.getListOfCategory()).hasSizeGreaterThan(1);
-            categories.getListOfCategory().forEach(h -> assertEquals("ListingStatus", h.getCategoryKey()));
-        } else {
-            assertEquals(NOT_FOUND.value(), response.getStatusCode());
-        }
-    }
-
-    @Test
-    @ToggleEnable(mapKey = MAP_KEY_CASE_FLAGS, withFeature = true)
-    @ExtendWith(FeatureToggleConditionExtension.class)
-    void shouldReturnBadRequestForRetrieveCaseFlagsByServiceIdWithAvailableExternalFlagIsEmpty() {
-        final var response = (ErrorResponse)
-            commonDataApiClient.retrieveCaseFlagsByServiceId(
-                HttpStatus.BAD_REQUEST,
-                "service-id=XXXX?available-external-flag="
-            );
-        assertEquals(response.getErrorCode(), 400);
-        assertEquals(response.getErrorDescription(), "Allowed values are Y or N");
-    }
 
 }
